Edition 1

+++ Assets/Advanced Multiplayer FPS (ForceCode)/Core/Scripts/Main Game/Scoreboard System/r_InGameScoreboardEntry.cs
@@ -48,7 +48,7 @@ namespace ForceCodeFPS
                 var transport = NetworkManager.Singleton.NetworkConfig.NetworkTransport;
                 if (transport is Unity.Netcode.Transports.UTP.UnityTransport utp)
                 {
-                    return utp.GetCurrentRtt(clientId);
+                    return (int)utp.GetCurrentRtt(clientId);
                 }
             }
             return 0;
